Conflicts:
	AndroidManifest.xml
	res/layout/emoticon_layout.xml
	res/values/strings.xml
	src/com/bsb/hike/HikeConstants.java
	src/com/bsb/hike/adapters/CentralTimelineAdapter.java
	src/com/bsb/hike/adapters/ConversationsAdapter.java
	src/com/bsb/hike/adapters/EmoticonAdapter.java
	src/com/bsb/hike/adapters/MessagesAdapter.java
	src/com/bsb/hike/adapters/StickerAdapter.java
	src/com/bsb/hike/adapters/StickerPageAdapter.java
	src/com/bsb/hike/filetransfer/FileTransferManager.java
	src/com/bsb/hike/models/Conversation.java
	src/com/bsb/hike/notifications/HikeNotification.java
	src/com/bsb/hike/notifications/ToastListener.java
	src/com/bsb/hike/ui/ChatThread.java
	src/com/bsb/hike/ui/ComposeChatActivity.java
	src/com/bsb/hike/ui/CreditsActivity.java
	src/com/bsb/hike/ui/FileSelectActivity.java
	src/com/bsb/hike/ui/GallerySelectionViewer.java
	src/com/bsb/hike/ui/HikeDialog.java
	src/com/bsb/hike/ui/HikeListActivity.java
	src/com/bsb/hike/ui/HikePreferences.java
	src/com/bsb/hike/ui/HomeActivity.java
	src/com/bsb/hike/ui/ProfileActivity.java
	src/com/bsb/hike/ui/SettingsActivity.java
	src/com/bsb/hike/ui/TellAFriend.java
	src/com/bsb/hike/ui/fragments/ConversationFragment.java
	src/com/bsb/hike/utils/HikeAppStateBaseFragmentActivity.java
	src/com/bsb/hike/utils/IntentManager.java
	src/com/bsb/hike/utils/StickerManager.java
	src/com/bsb/hike/utils/Utils.java

+++ src/com/bsb/hike/db/DBBackupRestore.java
@@ -8,19 +8,45 @@ import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.nio.channels.FileChannel;
+import java.util.Calendar;
+
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
 
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.os.Environment;
+import android.text.TextUtils;
 
 import com.bsb.hike.HikeConstants;
 import com.bsb.hike.HikeMessengerApp;
+import com.bsb.hike.analytics.AnalyticsConstants;
+import com.bsb.hike.analytics.HAManager;
+import com.bsb.hike.models.HikeAlarmManager;
 import com.bsb.hike.utils.CBCEncryption;
 import com.bsb.hike.utils.Logger;
+import com.bsb.hike.utils.Utils;
 
 public class DBBackupRestore
 {
-	private static DBBackupRestore _instance = null;
+	/**
+	 * @author gauravmittal
+	 * 	DBBackupRestore is a singleton class that performs are the backup/restore related
+	 * 	operations
+	 */
+	
+	public static final String RESTORE_EVENT_KEY = "rstr";
+
+	public static final String BACKUP_EVENT_KEY = "bck";
+
+	public static final String SIZE = "sz";
+
+	public static final String STATUS = "sts";
+
+	public static final String TIME_TAKEN = "tt";
+
+	private static volatile DBBackupRestore _instance = null;
 
 	private static final String HIKE_PACKAGE_NAME = "com.bsb.hike";
 
@@ -32,17 +58,19 @@ public class DBBackupRestore
 
 	private static final String[] resetTableNames = { DBConstants.STICKER_SHOP_TABLE, DBConstants.STICKER_CATEGORIES_TABLE };
 
-	private String backupToken;
-
-	private Context mContext;
+	private final Context mContext;
 
 	private DBBackupRestore(Context context)
 	{
 		this.mContext = context;
-		SharedPreferences settings = context.getSharedPreferences(HikeMessengerApp.ACCOUNT_SETTINGS, 0);
-		backupToken = settings.getString(HikeMessengerApp.BACKUP_TOKEN_SETTING, null);
 	}
 
+	/**
+	 * Gets the BDBackupRestore instance. Creates one it not already created.
+	 * @param context
+	 * @return
+	 * 		The BDBackupRestore instance
+	 */
 	public static DBBackupRestore getInstance(Context context)
 	{
 		if (_instance == null)
@@ -56,18 +84,49 @@ public class DBBackupRestore
 		return _instance;
 	}
 
+	/**
+	 * Schedules next auto backup.
+	 */
+	public void scheduleNextAutoBackup()
+	{
+		long scheduleTime = Utils.getTimeInMillis(Calendar.getInstance(), 3, 0, 0, 0);
+		// If the scheduled time is in the past OR the account restore process(at the time of signup) is not yet complete.
+		// Scheduled time is increased by 24 hours i.e. same time next day.
+		if (scheduleTime < System.currentTimeMillis() || !mContext.getSharedPreferences(HikeMessengerApp.ACCOUNT_SETTINGS, 0).getBoolean(HikeMessengerApp.RESTORE_ACCOUNT_SETTING, false))
+		{
+			scheduleTime += 24 * 60 * 60 * 1000;
+		}
+		HikeAlarmManager.setAlarm(mContext, scheduleTime, HikeAlarmManager.REQUESTCODE_PERIODIC_BACKUP, true);
+		Logger.d(getClass().getSimpleName(), "Scheduled next Auto-Backup for: " + Utils.getFormattedDateTimeFromTimestamp(scheduleTime/1000, mContext.getResources().getConfiguration().locale));
+	}
+
+	/**
+	 * Creates a complete backup of chats and the specified preferences.
+	 * @return
+	 * 	true for success, and false for for failure. 
+	 */
 	public boolean backupDB()
 	{
 		Long time = System.currentTimeMillis();
+		boolean result = true;
 		try
 		{
 			for (String fileName : dbNames)
 			{
 				File dbCopy = exportDatabse(fileName);
 				if (dbCopy == null || !dbCopy.exists())
-					return false;
+				{
+					result = false;
+					break;
+				}
 
 				File backup = getDBBackupFile(dbCopy.getName());
+				String backupToken = getBackupToken();
+				Logger.d(getClass().getSimpleName(), "encrypting with key: " + backupToken);
+				if (TextUtils.isEmpty(backupToken))
+				{
+					throw new Exception("Backup Token is empty");
+				}
 				CBCEncryption.encryptFile(dbCopy, backup, backupToken);
 				dbCopy.delete();
 			}
@@ -76,17 +135,28 @@ public class DBBackupRestore
 		{
 			deleteTempFiles();
 			e.printStackTrace();
-			return false;
+			result = false;
 		}
-		if (!updateBackupState())
+		if (result)
 		{
-			return false;
+			if (!updateBackupState(null))
+			{
+				result = false;
+			}
 		}
 		time = System.currentTimeMillis() - time;
-		Logger.d(getClass().getSimpleName(), "Backup complete!! in " + time / 1000 + "." + time % 1000 + "s");
-		return true;
+		Logger.d(getClass().getSimpleName(), "Backup " + result + " in " + time / 1000 + "." + time % 1000 + "s");
+		recordLog(BACKUP_EVENT_KEY,result,time);
+		return result;
 	}
 
+	/**
+	 * Creates a copy of the specified database of the application.
+	 * @param databaseName
+	 * 		The name of the database.
+	 * @return
+	 * 		The copy the database.
+	 */
 	public File exportDatabse(String databaseName)
 	{
 		Long time = System.currentTimeMillis();
@@ -122,42 +192,67 @@ public class DBBackupRestore
 		return dbCopy;
 	}
 
+	/**
+	 * Restores the complete backup of chats and the specified preferences.
+	 * @return
+	 * 	true for success, and false for for failure. 
+	 */
 	public boolean restoreDB()
 	{
 		Long time = System.currentTimeMillis();
+		boolean result = true;
 		BackupState state = getBackupState();
 		if (state == null)
 		{
-			return false;
+			result = false;
 		}
 		if (state.getDBVersion() > DBConstants.CONVERSATIONS_DATABASE_VERSION)
 		{
-			return false;
+			result = false;
 		}
-		try
+		if (result)
 		{
-			for (String fileName : dbNames)
+			try
 			{
-				File currentDB = getCurrentDBFile(fileName);
-				File dbCopy = getDBCopyFile(currentDB.getName());
-				File backup = getDBBackupFile(dbCopy.getName());
-				CBCEncryption.decryptFile(backup, dbCopy, backupToken);
-				importDatabase(dbCopy);
-				dbCopy.delete();
+				for (String fileName : dbNames)
+				{
+					File currentDB = getCurrentDBFile(fileName);
+					File dbCopy = getDBCopyFile(currentDB.getName());
+					File backup = getDBBackupFile(dbCopy.getName());
+					String backupToken = getBackupToken();
+					Logger.d(getClass().getSimpleName(), "decrypting with key: " + backupToken);
+					if (TextUtils.isEmpty(backupToken))
+					{
+						throw new Exception("Backup Token is empty");
+					}
+					CBCEncryption.decryptFile(backup, dbCopy, backupToken);
+					importDatabase(dbCopy);
+					dbCopy.delete();
+				}
+			}
+			catch (Exception e)
+			{
+				deleteTempFiles();
+				e.printStackTrace();
+				result = false;
 			}
 		}
-		catch (Exception e)
+		if (result)
 		{
-			deleteTempFiles();
-			e.printStackTrace();
-			return false;
+			state.restorePrefs(mContext);
+			postRestoreSetup(state);
 		}
-		postRestoreSetup(state);
 		time = System.currentTimeMillis() - time;
-		Logger.d(getClass().getSimpleName(), "Restore complete!! in " + time / 1000 + "." + time % 1000 + "s");
-		return true;
+		Logger.d(getClass().getSimpleName(), "Restore " + result + " in " + time / 1000 + "." + time % 1000 + "s");
+		recordLog(RESTORE_EVENT_KEY,result,time);
+		return result;
 	}
 
+	/**
+	 * Replaces the current Application database file with the provided database file
+	 * @param dbCopy
+	 * 		The file to placed as the new database.
+	 */
 	private void importDatabase(File dbCopy)
 	{
 		Long time = System.currentTimeMillis();
@@ -189,6 +284,39 @@ public class DBBackupRestore
 		time = System.currentTimeMillis() - time;
 		Logger.d(getClass().getSimpleName(), "DB import complete!! in " + time / 1000 + "." + time % 1000 + "s");
 	}
+	
+	private void recordLog(String eventKey, boolean result, long timeTaken)
+	{
+		JSONObject metadata = new JSONObject();
+		try
+		{
+			JSONArray sizes = new JSONArray();
+			for (String fileName : dbNames)
+			{
+				File currentDB = getCurrentDBFile(fileName);
+				File dbCopy = getDBCopyFile(currentDB.getName());
+				File backup = getDBBackupFile(dbCopy.getName());
+				sizes.put(backup.length());
+			}
+			metadata
+			.put(HikeConstants.EVENT_KEY, eventKey)
+			.put(SIZE, sizes)
+			.put(STATUS, result)
+			.put(TIME_TAKEN, timeTaken);
+			HAManager.getInstance().record(AnalyticsConstants.NON_UI_EVENT, AnalyticsConstants.ANALYTICS_BACKUP, metadata);
+		}
+		catch(JSONException e)
+		{
+			Logger.d(AnalyticsConstants.ANALYTICS_TAG, "invalid json");
+		}
+	}
+	
+	private String getBackupToken()
+	{
+		SharedPreferences settings = mContext.getSharedPreferences(HikeMessengerApp.ACCOUNT_SETTINGS, 0);
+		String backupToken = settings.getString(HikeMessengerApp.BACKUP_TOKEN_SETTING, null);
+		return backupToken;
+	}
 
 	private void postRestoreSetup(BackupState state)
 	{
@@ -200,8 +328,14 @@ public class DBBackupRestore
 		{
 			HikeConversationsDatabase.getInstance().clearTable(table);
 		}
+		HikeConversationsDatabase.getInstance().upgradeForStickerShopVersion1();
 	}
 
+	/**
+	 * Closes the closeables.
+	 * @param closeables
+	 * 		Set of the closeables to be closed.
+	 */
 	private void closeChannelsAndStreams(Closeable... closeables)
 	{
 		for (Closeable closeable : closeables)
@@ -218,6 +352,11 @@ public class DBBackupRestore
 		}
 	}
 
+	/**
+	 * Checks if a backup is there or not.
+	 * @return
+	 * 		true is the backup is available, false otherwise.
+	 */
 	public boolean isBackupAvailable()
 	{
 		for (String fileName : dbNames)
@@ -228,11 +367,15 @@ public class DBBackupRestore
 			if (!backup.exists())
 				return false;
 		}
-		if(!getBackupStateFile().exists())
+		BackupState state = getBackupState();
+		if (state != null)
 		{
-			return false;
+			if (state.getBackupTime() > 0)
+			{
+				return true;
+			}
 		}
-		return true;
+		return false;
 	}
 
 	public long getLastBackupTime()
@@ -245,6 +388,9 @@ public class DBBackupRestore
 		return -1;
 	}
 
+	/**
+	 * Deletes the temporary files.
+	 */
 	private void deleteTempFiles()
 	{
 		for (String fileName : dbNames)
@@ -255,6 +401,9 @@ public class DBBackupRestore
 		}
 	}
 
+	/**
+	 * Deletes all the backup files.
+	 */
 	public void deleteAllFiles()
 	{
 		for (String fileName : dbNames)
@@ -265,6 +414,8 @@ public class DBBackupRestore
 			dbCopy.delete();
 			backup.delete();
 		}
+		getBackupStateFile().delete();
+		deleteTempFiles();
 	}
 
 	private File getCurrentDBFile(String dbName)
@@ -292,10 +443,21 @@ public class DBBackupRestore
 		new File(HikeConstants.HIKE_BACKUP_DIRECTORY_ROOT).mkdirs();
 		return new File(HikeConstants.HIKE_BACKUP_DIRECTORY_ROOT, BACKUP);
 	}
-	
-	private boolean updateBackupState()
+
+	/**
+	 * Update the backup state file
+	 * @param state
+	 * 
+	 * @return
+	 * 		Success or failure
+	 */
+	private boolean updateBackupState(BackupState state)
 	{
-		BackupState state = new BackupState(dbNames[0], DBConstants.CONVERSATIONS_DATABASE_VERSION);
+		if (state == null)
+		{
+			state = new BackupState(dbNames[0], DBConstants.CONVERSATIONS_DATABASE_VERSION);
+			state.backupPrefs(mContext);
+		}
 		File backupStateFile = getBackupStateFile();
 		FileOutputStream fileOut = null;
 		ObjectOutputStream out = null;
@@ -350,6 +512,22 @@ public class DBBackupRestore
 			closeChannelsAndStreams(fileIn,in);
 		}
 		return state;
+	}
+
+	/**
+	 * Takes fresh backup of the preferences.
+	 * @return
+	 * 		The success or failure.
+	 */
+	public boolean updatePrefs()
+	{
+		BackupState state = getBackupState();
+		if (state == null)
+		{
+			state = new BackupState(dbNames[0], DBConstants.CONVERSATIONS_DATABASE_VERSION);
+		}
+		state.backupPrefs(mContext);
+		return updateBackupState(state);
 	}
 
 }
